pyproject.toml
[project]
name = "mumzworld-email-triage"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "faiss-cpu>=1.7.0",
    "pydantic>=2.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=4.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
